Addressed first round of review comments
